Add Window class that handles GLFW side
Would be best to keep both GLFW and OpenGL in their own classes in case there would be an interest in implementing different libraries e.g. Vulkan. For now rendering is all over the place, but plan is to have some sort of Render API class and possibly add command pattern to that